Number Recognition/requirements.txt
numpy==1.17.0
Flask==1.0.2
tensorflow==1.15.0
Pillow==7.1.1
gunicorn